moved the anim.ts into the utils folderand changed the width of the navbar mobile view in to 100%
